5.8.7 add support seley

+++ kivy_ios/recipes/pyaes/__init__.py
@@ -20,11 +20,21 @@ class pyAESRecipe(CythonRecipe):
     url = "https://github.com/XX-net/pyAES/archive/{version}.zip"
     depends = ["python"]
     libraries = [
-        # "libcryptopp.a",
+        "libpyaes.a",
+        "libcryptopp.a",
+        "libaes_cfb.a",
     ]
     cryptopp_version = "8.9"
 
+    def download(self):
+        pass
+
+    def extract(self):
+        pass
+
     def extract_arch(self, arch):
+        # Copy pyAES from parent dir.
+
         def ignore(src, names):
             # print(src, names)
             if src.endswith("pyAES") and "build" in names:
@@ -32,12 +42,16 @@ class pyAESRecipe(CythonRecipe):
             else:
                 return set()
 
-        build_dir = join(self.ctx.build_dir, self.name, arch)
-        dest_dir = join(build_dir, self.archive_root)
-        shutil.rmtree(dest_dir, ignore_errors=True)
-        shutil.copytree(pyAES_path, dest_dir, ignore=ignore)
+        logger.debug("pyAES path:%s", pyAES_path)
+        if os.path.isdir(pyAES_path):
+            build_dir = join(self.ctx.build_dir, self.name, arch)
+            dest_dir = join(build_dir, self.archive_root)
+            shutil.rmtree(dest_dir, ignore_errors=True)
+            shutil.copytree(pyAES_path, dest_dir, ignore=ignore)
 
     def prebuild_arch(self, arch):
+        logger.debug("pyaes prebuild %s", arch)
+        # download cryptopp if not downloaded.
         build_dir = self.get_build_dir(arch.arch)
         if not os.path.isfile(os.path.join(build_dir, "cryptopp", "aes.h")):
             self.download_file(f"https://github.com/XX-net/cryptopp/archive/refs/tags/{self.cryptopp_version}.zip", "cryptopp.zip", build_dir)
@@ -45,75 +59,75 @@ class pyAESRecipe(CythonRecipe):
             shutil.rmtree(os.path.join(build_dir, "cryptopp"))
             os.rename(os.path.join(build_dir, f"cryptopp-{self.cryptopp_version}"), os.path.join(build_dir, "cryptopp"))
 
+    def build_cryptopp(self, arch):
+        build_dir = self.get_build_dir(arch.arch)
+        chdir(build_dir)
+
+        os.environ["IOS_SDK"] = "iPhoneOS"
+
+        if arch.arch in ["x86_64"]:
+            os.environ["IOS_CPU"] = "x86_64"
+        else:
+            os.environ["IOS_CPU"] = "arm64"
+        # IOS_SDK=iPhoneOS IOS_CPU=arm64 source TestScripts/setenv-ios.sh
+        # make
+
+        build_ios_lib = sh.Command(os.path.join(build_dir, "build_cryptopp_ios.sh"))
+        shprint(build_ios_lib)
+
+    def build_aes_cfb(self, arch):
+        # cmake -G Xcode -Bbuild -DCMAKE_TOOLCHAIN_FILE=ios.toolchain.cmake  -DPLATFORM=OS64 -DCMAKE_SYSTEM_NAME=iOS -DCMAKE_Swift_COMPILER_FORCED=true -DCMAKE_OSX_DEPLOYMENT_TARGET=12.0
+        cmake = sh.Command("/usr/local/bin/cmake")
+        if arch.arch in ["x86_64"]:
+            shprint(cmake, "-G", "Xcode", "-Bbuild_aes", "-DCMAKE_TOOLCHAIN_FILE=ios.toolchain.cmake",
+                    "-DPLATFORM=SIMULATOR64",
+                    )
+        else:
+            shprint(cmake, "-G", "Xcode", "-Bbuild_aes", "-DCMAKE_TOOLCHAIN_FILE=ios.toolchain.cmake",
+                    "-DPLATFORM=OS64",
+                    )
+
+        # cmake --build build -v -j8 --config Release --target aes_cfb
+        shprint(cmake, "--build", "build_aes", "--config", "Release", "--target", "aes_cfb")
+
     def build_arch(self, arch):
-        # zsh = sh.Command("/bin/zsh")
-        # shprint(zsh)
-        # os.system("zsh")
+        build_dir = self.get_build_dir(arch.arch)
+        logger.info("Building pyAES {} in {}".format(arch.arch, build_dir))
+        chdir(build_dir)
+
+        os.environ["PATH"] = "/opt/local/bin:/opt/local/sbin:/usr/local/bin:/System/Cryptexes/App/usr/bin:/usr/bin:/bin:/usr/sbin:/sbin:/usr/local/go/bin:/Library/Apple/usr/bin"
 
         super().build_arch(arch)
-        # build_dir = self.get_build_dir(arch.arch)
-        # logger.info("Building pyAES {} in {}".format(arch.arch, build_dir))
-        # chdir(build_dir)
-        # 
-        # os.environ["PATH"] = "/opt/local/bin:/opt/local/sbin:/usr/local/bin:/System/Cryptexes/App/usr/bin:/usr/bin:/bin:/usr/sbin:/sbin:/usr/local/go/bin:/Library/Apple/usr/bin"
-        # 
-        # cmake = sh.Command("/usr/local/bin/make")
-        # if arch.arch in ["x86_64"]:
-        #     shprint(cmake, "-G", "Xcode", "-Bbuild", "-DCMAKE_TOOLCHAIN_FILE=ios.toolchain.cmake",
-        #             "-DPLATFORM=SIMULATOR64",
-        #             '-DBORINGSSL_PREFIX=BSSL',
-        #             "-DBORINGSSL_PREFIX_SYMBOLS=" + symbols_fp,
-        #             )
-        # else:
-        #     shprint(cmake, "-G", "Xcode", "-Bbuild", "-DCMAKE_TOOLCHAIN_FILE=ios.toolchain.cmake",
-        #             "-DPLATFORM=OS64",
-        #             '-DBORINGSSL_PREFIX=BSSL',
-        #             "-DBORINGSSL_PREFIX_SYMBOLS=" + symbols_fp,
-        #             )
-        # 
-        # build_ios_lib = sh.Command(os.path.join(build_dir, "build_ios_lib.zsh"))
-        # shprint(build_ios_lib)
-        # print("build " + arch.arch)
+
+        self.build_cryptopp(arch)
+
+        self.build_aes_cfb(arch)
 
     def postbuild_arch(self, arch):
-        super().postbuild_arch(arch)
-    #     build_dir = self.get_build_dir(arch.arch)
-    #
-    #     libraries = [
-    #         "build/boringssl/ssl/Release-iphoneos/libssl.a",
-    #         "build/boringssl/crypto/Release-iphoneos/libcrypto.a",
-    #         "build/boringssl/decrepit/Release-iphoneos/libdecrepit.a",
-    #
-    #         "build/brotli/Release-iphoneos/libbrotlicommon.a",
-    #         "build/brotli/Release-iphoneos/libbrotlidec.a",
-    #
-    #         "build/cert_decompress/Release-iphoneos/libcert_decompress.a",
-    #         "build/getpeercert/Release-iphoneos/libgetpeercert.a",
-    #         "build/cffi_boringssl/Release-iphoneos/libcffi_boringssl.a",
-    #     ]
-    #
-    #     for fp in libraries:
-    #         fpl = fp.split("/")
-    #         if arch.arch == "x86_64":
-    #             fpl[-2] = "Release-iphonesimulator"
-    #         else:
-    #             fpl[-2] = "Release-iphoneos"
-    #
-    #         fp = "/".join(fpl)
-    #         src = os.path.join(build_dir, fp)
-    #
-    #         fn = fpl[-1]
-    #         if fn == "libssl.a":
-    #             fn = "libbssl.a"
-    #         elif fn == "libcrypto.a":
-    #             fn = "libbcrypto.a"
-    #         dst = os.path.join(build_dir, fn)
-    #         shutil.copyfile(src, dst)
+        build_dir = self.get_build_dir(arch.arch)
 
-    # def install(self):
-    #     dst = os.path.join(self.ctx.site_packages_dir, "boringssl.so")
-    #     with open(dst, "w") as fd:
-    #         fd.write(" ")
+        shutil.copyfile(
+            join(build_dir, "cryptopp", "libcryptopp.a"),
+            join(build_dir, "libcryptopp.a")
+            )
+
+        if arch.arch in ["x86_64"]:
+            shutil.copyfile(
+                join(build_dir, "build_aes", "aes_cfb", "Release-iphonesimulator", "libaes_cfb.a"),
+                join(build_dir, "libaes_cfb.a")
+            )
+        else:
+            shutil.copyfile(
+                join(build_dir, "build_aes", "aes_cfb", "Release-iphoneos", "libaes_cfb.a"),
+                join(build_dir, "libaes_cfb.a")
+            )
+        super().postbuild_arch(arch)
+
+    def install(self):
+        # put an empty so file
+        dst = os.path.join(self.ctx.site_packages_dir, "py_aes_cfb.so")
+        with open(dst, "w") as fd:
+            fd.write(" ")
 
 
 recipe = pyAESRecipe()
